Move the original source.tar.gz files into the homepage repo.
I performed a due diligence exercise to validate that these are indeed
the correct and unadultered files by trawling the FreeBSD and Gentoo
repositories for evindence of original SHA256 hash values.

This information is now recorded in the per release files.

I could not find evidence for the 4.0.4 tarball sha256, but
it is included anyway as I have no reason to suspect it.

For the 1.1 tarball, the sha256 does not match what FreeBSD
ports expected it to be, so I have not added it at this time.