Create vmod_51degrees_device_detection.json
Hi Varnish team. We are a company that provide access to real time digital data services and we have been developing and supporting fast and accurate device detection technologies with integrations across different platforms and languages. We have been supporting our Device Detection engine with Varnish for a while and we would like our module to be listed in the VMODs list please. Kind Regards.