Generic authorship and Copyright
as of today, each page on https://varnish-cache.org/ redendered from the
sphinx source contains the line

	©2016-2023, Poul-Henning Kamp

While this is true for much of the content, it is not for all of it and
because we, several people from UPLEX, intend to contribute significant
amounts of documentation in the near future, we would like to change
this into something more generic, without the need to state specific
copyright and authorship on each page.

This is a suggestion only and we are open to alternative proposals. And
at any rate, no-one intends to lessen phk's contribution.